always load document from disk
- read documents directly from disk, rather than db, to sidestep desync issues
