+++ src/preload/client/importer/frontmatter.ts
@@ -266,3 +266,54 @@ function preprocessRawFrontMatter(content: string) {
       })
   );
 }
+
+function preprocessChroniclesFrontMatter(content: string) {
+  // Regular expression to match key-value pairs in front matter
+  return content
+    .replace(/^(\w+):\s*$/gm, '$1: ""') // Handle keys with no values
+    .replace(/^(\w+):\s*(.+)$/gm, (match, key, value) => {
+      // Check if value contains special characters that need quoting
+      if (value.match(/[:{}[\],&*#?|\-<>=!%@`]/) || value.includes("\n")) {
+        // If the value is not already quoted, wrap it in double quotes
+        if (!/^['"].*['"]$/.test(value)) {
+          // Escape any existing double quotes in the value
+          value = value.replace(/"/g, '\\"');
+          return `${key}: "${value}"`;
+        }
+      }
+      return match; // Return unchanged if no special characters
+    });
+}
+
+// naive frontmatter parser for files formatted in chronicles style...
+// which  just means a regular markdown file + yaml front matter
+// ... todo: use remark ecosystem parser
+export function parseChroniclesFrontMatter(content: string) {
+  // Regular expression to match front matter (--- at the beginning and end)
+  const frontMatterRegex = /^---\n([\s\S]*?)\n---\n*/;
+
+  // Match the front matter
+  const match = content.match(frontMatterRegex);
+  if (!match) {
+    return {
+      frontMatter: {}, // No front matter found
+      body: content, // Original content without changes
+    };
+  }
+
+  // Extract front matter and body
+  const frontMatterContent = preprocessChroniclesFrontMatter(match[1]);
+  const body = content.slice(match[0].length); // Content without front matter
+
+  // Parse the front matter using yaml
+  const frontMatter = yaml.parse(frontMatterContent);
+  frontMatter.tags = frontMatter.tags
+    .split(",")
+    .map((tag: string) => tag.trim())
+    .filter(Boolean);
+
+  return {
+    frontMatter,
+    body,
+  };
+}
